+++ src/Listener/ProductListener.php
@@ -23,8 +23,8 @@ class ProductListener
     use UserTrait;
 
     public function __construct(
-        private Security $security,
-        private EntityManagerInterface $entityManager
+        private readonly Security $security,
+        private readonly EntityManagerInterface $entityManager
     ) {
     }