+++ src/Service/BasketServiceInterface.php
@@ -24,17 +24,21 @@ interface BasketServiceInterface
 
     public function deleteProductItem(Request $request): array;
 
+    public function generateSecurityToken(): string;
+
     public function get(): ?Basket;
 
     public function getJson(): array;
 
     public function processStripePayment($session): void;
 
+    public function itemsShipped(string $number): Basket;
+
     public function updateTotals(): void;
 
     public function validate(Request $request): string;
 
-    public function validated(Basket $basket): void;
+    public function paid(Basket $basket): void;
 }